hfpclient: Add a script to emulate an HF
This is the opposite of the HFP "server" (HFAG).
In this script, we emulate a car kit (HF) and have the phone/modem
connect to us. We can then send it AT commands.

We are a "client" in the sense the HFAG (phone/modem) gets AT commands
from us, processes the commands, and sends us responses.

However, this acts as a server in the RFCOMM sense since the HFAG
is supposed to initiate a connection to the HF.